+++ src/engine.rs
@@ -3,3 +3,9 @@ pub mod port_buffer;
 pub mod regressor;
 pub mod graph;
 pub mod optimizer;
+pub mod hogwild;
+pub mod persistence;
+pub mod quantization;
+pub mod buffer_handler;
+pub mod multithread_helpers;
+pub mod serving;
