+++ src/engine/block/iterators.rs
@@ -1,6 +1,6 @@
 use crate::engine::optimizer::OptimizerTrait;
 
-use crate::feature_buffer;
+use crate::namespace::feature_buffer;
 use crate::engine::port_buffer;
 use crate::engine::regressor::{BlockCache, BlockTrait};
