+++ src/engine/persistence.rs
@@ -9,9 +9,9 @@ use std::io::Read;
 
 use crate::model_instance;
 use crate::engine::regressor;
-use crate::vwmap;
+use crate::namespace::vwmap;
 
-use crate::multithread_helpers::BoxedRegressorTrait;
+use crate::engine::multithread_helpers::BoxedRegressorTrait;
 use crate::engine::regressor::Regressor;
 
 const REGRESSOR_HEADER_MAGIC_STRING: &[u8; 4] = b"FWRE"; // Fwumious Wabbit REgressor
@@ -207,8 +207,8 @@ mod tests {
     // Note this useful idiom: importing names from outer (for mod tests) scope.
     use super::*;
     use crate::assert_epsilon;
-    use crate::feature_buffer;
-    use crate::feature_buffer::{HashAndValue, HashAndValueAndSeq};
+    use crate::namespace::feature_buffer;
+    use crate::namespace::feature_buffer::{HashAndValue, HashAndValueAndSeq};
     use crate::model_instance::Optimizer;
     use crate::engine::optimizer;
     use crate::engine::optimizer::OptimizerTrait;
